+++ src/main/java/org/truffleruby/language/dispatch/CachedDispatchNode.java
@@ -101,13 +101,6 @@ public abstract class CachedDispatchNode extends DispatchNode {
     protected void applySplittingInliningStrategy(DirectCallNode callNode, InternalMethod method) {
         if (callNode.isCallTargetCloningAllowed() && method.getSharedMethodInfo().shouldAlwaysClone()) {
             callNode.cloneCallTarget();
-        } else if (getContext().getOptions().CALL_WITH_BLOCK_ALWAYS_CLONE) {
-            final RubyCallNode rubyCallNode = findRubyCallNode();
-            if (rubyCallNode != null && rubyCallNode.hasLiteralBlock()) {
-                // If the call has a literal block, it is interesting to split so the method can
-                // specialize for that block only and not all blocks given to the method so far.
-                callNode.cloneCallTarget();
-            }
         }
 
         if (sendingFrames() && getContext().getOptions().INLINE_NEEDS_CALLER_FRAME && callNode.isInlinable()) {